refactor: pass env vars as ConfigOptions to execute
